feat: 문서 추가 및 API 정의 #5
