+++ stock-api.yml
@@ -6,47 +6,47 @@ info:
 servers:
 - url: http://localhost:8080/stocks
 tags:
-- name: 저장된 주식 반환 API
-  description: 저장된 5일치의 주식 데이터를 날짜 기준 오름차순으로 반환합니다.
+- name: stockController
+  description: 주식 5일치 API
 paths:
-  /stocks/:
+  /stocks:
     get:
       tags:
-      - 저장된 주식 반환 API
+      - stockController
       description: 주식 데이터 5일치 반환
       operationId: getFinanceInfo
       responses:
         "200":
           description: 주식 데이터 5일치 반환 성공
           content:
-            '*/*':
+            application/json:
               schema:
-                $ref: '#/components/schemas/ApiResponseDto'
-  /stocks:
-    get:
-      tags:
-      - 저장된 주식 반환 API
-      description: 주식 데이터 5일치 반환
-      operationId: getFinanceInfo_1
-      responses:
-        "200":
-          description: 주식 데이터 5일치 반환 성공
-          content:
-            '*/*':
-              schema:
-                $ref: '#/components/schemas/ApiResponseDto'
+                type: array
+                items:
+                  $ref: '#/components/schemas/stock'
 components:
   schemas:
-    ApiResponseDto:
+    stock:
       type: object
       properties:
-        code:
-          type: string
-          enum:
-          - SUCCESS
-          - ERROR
-          - BAD_REQUEST
-        detail:
-          type: string
-        data:
-          type: object
+        date:
+          description: 주식의 날짜
+          example: 2023-02-17
+        name:
+          description: 주식의 이름
+          example: 005930.KS
+        high:
+          description: 주식 최고 가격
+          example: 63300
+        low:
+          description: 주식 최저 가격
+          example: 62400
+        open:
+          description: 시장 오픈 가격
+          example: 62900
+        close:
+          description: 시장 마감 가격
+          example: 62600
+        volume:
+          description: 거래량
+          example: 28613
